Revert "Start allowing larger numeric types for reads"
This reverts commit e4a0a08d136c14009c76ce9187287de0ae5bf5c0.

+++ SabreTools.IO/Extensions/ByteArrayExtensions.cs
@@ -504,7 +504,7 @@ namespace SabreTools.IO.Extensions
             var sb = new StringBuilder();
 
             // Check for strings
-            long offset = 0;
+            int offset = 0;
             while (offset < bytes.Length)
             {
                 // Read the next character from the stream
